moving getURLHash() to utils to share
